+++ platforms/esp32/apps/memfault_demo_app/main/console_example_main.c
@@ -130,7 +130,6 @@ void app_main()
   esp_console_register_help_command();
   register_system();
   register_wifi();
-  memfault_register_cli();
 
   /* Prompt to be printed before each line.
    * This can be customized, made dynamic, etc.